requirements.txt
asgiref==3.11.0
django==4.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
pillow==12.0.0
pyjwt==2.10.1
sqlparse==0.5.5
